+++ tests/Type/GeneralTypeTest.php
@@ -29,6 +29,7 @@ class GeneralTypeTest extends TypeInferenceTestCase
         yield from $this->gatherAssertTypes(__DIR__.'/data/optional-helper.php');
         yield from $this->gatherAssertTypes(__DIR__.'/data/abstract-manager.php');
         yield from $this->gatherAssertTypes(__DIR__.'/data/facades.php');
+        yield from $this->gatherAssertTypes(__DIR__.'/data/where-relation.php');
     }
 
     /**